+++ SmallBeachFarm/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* Added support for other mods patching the custom tilesheet.
 * Fixed missing warp tile.
 
 ## 1.4.3